Start new tasks immediately on startup
Previously the runtime would wait for an event from the database before
it checked if there were jobs for it to run. This doesn't work so well
when there are tasks already queued up in the database.

This commit changes it to check for new tasks immediately on startup before
entering the main event loop. That way pre-existing tasks are launched
immediately if available.