Prevent workers from deleting themselves from the database
This should never really happen in production but it's possible that a
heartbeat hasn't executed for a while in development. (e.g. if running
on a laptop that went to sleep for a while).

In this case, if nobody else has removed the worker from the database
then there is really no reason for it to remove itself from the
database.