Make the shutdown state of WorkerHandle persistent (#26)
Currently, Worker::run resets the state of its WorkerHandle when it is
first called. This results in a problematic race condition between the
handle being created and Worker::run being called. If the handle is
shutdown during that time then the shutdown signal will be lost. A
similar problem happens if Worker::run is run in an error retry loop.

The easy way to avoid this is to make the WorkerHandle state persistent.
If shutdown() is called, then it will stay in that state until
explicitly cleared by the user of the library. This avoids pretty much
all race conditions unless the user of the library needs to use reset,
in which case they can synchronize things themselves.