Explicitly override the Cache-Control header within the registry
By default, GCS will return a Cache-Control header with max-age=3600.
This is fine for the .crate files in the registry but causes issues for
the index files which are updated when a new crate version is published.

By setting the cache policy to no-cache for the index files we can
ensure that we don't have to worry about old cached responses.