fix: Make Task::follow_logs immediately return existing log messages
Previously this would wait for an initial event before doing anything.
This is wrong, since if it is run then it won't print anything until an
event is emitted, which could take a long time to occur.

This commit changes it to immediately fetch log messages and only wait
for events after it has yielded those.