Automatically set SQLX_OFFLINE when crates are used as dependencies
Users of these libraries may not have a database with the right schema.
We don't want to require that so we need to make sure to use the saved
.sqlx queries when building as a package.

The way this is done is by setting SQLX_OFFLINE=true and
SQLX_OFFLINE_DIR=.sqlx in the build.rs scripts of the relevant packages.
This means that the DATABASE_URL environment variable will be ignored
and that `cargo sqlx prepare` will not prepare queries for these
libraries.

To make it work locally, if DURABLE_DEVELOPMENT=1 then we don't do any
overrides. We then set DURABLE_DEVELOPMENT=1 in .cargo/config.toml.