Explicitly handle some recoverable sqlx errors instead of failing
The whole goal of durable is for its workflows to be reliable. While we
can't handle some errors, there are a subset of sqlx errors that are not
the fault of the workflow itself and should instead be handled by
performing a retry.

This commit handles these and translates them into either retries or
runtime-level errors.