Fix trigger definition for when a task is notified after being suspended (#55)
The trigger was only firing when running_on is changed to NULL.
Unfortunately, the trigger reviving the suspended task actually sets
running_on to the id of a random worker. So no notifications are sent
and the task never starts executing.

This PR changes the trigger to fire whenever the task state changes from
a non-active state to an active one.