Explicitly reconnect PgListener before returning Event::Lagged
By default, PgListener reconnects on the next call to try_recv. However,
reading the database state in order to recover from missing a message
happens before that call to try_recv. This means that there is a race
condition. If a notification is emitted after the recovery but before
the connection is established then it will be lost.

In practice, this manifests as the worker getting stuck in some rare
conditions where some tasks are marked as ready during the race period
and then the worker remains stuck until another task becomes ready,
potentially forever. It is much easier to trigger this when the pool is
under high load, since queueing within sqlx::PgPool means that there is
a much larger period between recovering and the new listener connection
being ready.

This commit fixes the issue by running a dummy query before returning
Event::Lagged, which forces PgListener to reconnect before we perform
recovery.