Speed up task resumption by caching a list of events (#75)
When a task is resumed it needs to replay up to the point where it left
off. This involves quite a few database queries and can be rather slow.
By reading the whole list of events at the start we can avoid hitting
the database during replay, which makes things much faster.